move graphql resolvers and schema to a separate folder
